+++ nzme_skynet/core/controls/baseelement.py
@@ -5,9 +5,13 @@ from nzme_skynet.core.controls import highlight_state
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.common.action_chains import ActionChains
+from selenium.webdriver.remote.webelement import WebElement
+from selenium.webdriver.common.by import By
 import time
 import logging
 from nzme_skynet.core.utils.log import Logger
+from typing import Union
+from selenium.webdriver.phantomjs.webdriver import WebDriver
 
 
 Logger.configure_logging()
@@ -29,6 +33,7 @@ class BaseElement(object):
         self._locator = locator
 
     def _find_element(self):
+        # type: () -> WebElement
         """
         This methods validates the presence (but not visibility) of the element on the DOM within a DEFAULT_TIMEOUT
         period of 5 seconds. This method returns the web-element in successful case and logs an exception when
@@ -37,8 +42,9 @@ class BaseElement(object):
         :return: web element
         """
         try:
-            return WebDriverWait(self.driver, DefaultTimeouts.DEFAULT_TIMEOUT).until(ec.presence_of_element_located((self._by, self._locator)))
-        except Exception as e:
+            return WebDriverWait(self.driver, DefaultTimeouts.DEFAULT_TIMEOUT).until\
+                (ec.presence_of_element_located((self._by, self._locator)))
+        except Exception:
             logger.exception("Timeout: Failed to find element {0}".format(self._locator))
             raise
 
@@ -55,7 +61,14 @@ class BaseElement(object):
         return self._find_element().find_elements(by, locator)
 
     @property
+    def parent(self):
+        # type: () -> WebElement
+        # TODO: return  BaseElement object
+        return self._find_element().find_element(By.XPATH, "..")
+
+    @property
     def driver(self):
+        # type: () -> Union[WebDriver, None]
         """
         This method calls Selenium's DriverFactory to build driver and returns the driver to be used in the tests
         :return: driver
@@ -89,7 +102,7 @@ class BaseElement(object):
         try:
             self._find_element()
             return True
-        except Exception as e:
+        except Exception:
             logger.debug("Element {0} is not present in the DOM".format(self._locator))
             return False
 
@@ -102,7 +115,7 @@ class BaseElement(object):
         :return: element or False
         """
         if highlight_state():
-            elem = self.will_be_visible()
+            elem = self._find_element()
 
             def apply_style(style):
                 self.driver.execute_script("arguments[0].setAttribute('style', arguments[1]);",
@@ -158,6 +171,7 @@ class BaseElement(object):
     # Visibility, Presence, Clickability
 
     def is_currently_visible(self, time=DefaultTimeouts.SHORT_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will instantly (1 second) validate if the web element is is not only displayed but also has a
         height and width that is greater than 0. Returns the element if it is present. Returns False and
@@ -169,6 +183,7 @@ class BaseElement(object):
         return self.will_be_visible(time=time)
 
     def will_be_visible(self, time=DefaultTimeouts.LARGE_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will wait for 10 seconds to validate if the web element is is not only displayed but also has a
         height and width that is greater than 0. Returns the element if it is present. Returns False and
@@ -184,6 +199,7 @@ class BaseElement(object):
             return False
 
     def is_currently_present(self, time=DefaultTimeouts.SHORT_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will instantly (1 second) validate if the web element is present on the DOM of a page. This does
         not guarantee the visibility of the element though. Returns the element if it is present. Returns False and
@@ -195,6 +211,7 @@ class BaseElement(object):
         return self.will_be_present(time=time)
 
     def will_be_present(self, time=DefaultTimeouts.LARGE_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will wait for 10 seconds to validate if the web element is present on the DOM of a page. This does
         not guarantee the visibility of the element though. Returns the element if it is present. Returns False and
@@ -210,6 +227,7 @@ class BaseElement(object):
             return False
 
     def is_not_displayed(self, time=DefaultTimeouts.SHORT_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will instantly (1 second) look for a web element that is either invisible or not
         present on the DOM. Returns True if the element is invisible and False when visible.
@@ -220,6 +238,7 @@ class BaseElement(object):
         return self.will_not_be_displayed(time=time)
 
     def will_not_be_displayed(self, time=DefaultTimeouts.LARGE_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will wait for 10 seconds to validate if the web element is either invisible or not
         present on the DOM. Returns True if the element is invisible and False when visible.
@@ -234,6 +253,7 @@ class BaseElement(object):
             return False
 
     def is_ready_to_interact(self, time=DefaultTimeouts.SHORT_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will instantly (1 second) look for a web element to enabled and be visible.
 
@@ -246,8 +266,8 @@ class BaseElement(object):
         """
         return self.will_be_ready_to_interact(time=time)
 
-
     def will_be_ready_to_interact(self, time=DefaultTimeouts.LARGE_TIMEOUT):
+        # type: (int) -> Union[WebElement, bool]
         """
         This method will wait for a web element to enabled and be visible within 10 seconds.
 
@@ -272,8 +292,6 @@ class BaseElement(object):
         hover.perform()
 
     def focus(self):
-        """
-        TODO: Implement focus function
-        :return:
-        """
-        raise NotImplementedError
+        hover = ActionChains(self.driver).move_to_element(self._find_element())
+        hover.click()
+        hover.perform()