+++ nzme_skynet/core/controls/textinput.py
@@ -21,6 +21,7 @@ class TextInput(Text):
         :return: clear or False
         """
         self._highlight()
+        self.focus()
         self._find_element().clear()
 
     @property
@@ -44,3 +45,13 @@ class TextInput(Text):
         self.is_currently_visible()
         self.clear()
         self._find_element().send_keys(value)
+
+    def update_value(self, new_value):
+        """
+        Completely updates the exiting text value to a new value specified
+        :param new_value: the new value to update
+        """
+        # Check if text element is visible
+        # Clear existing value
+        # Set new value
+        raise NotImplemented