+++ nzme_skynet/core/driver/mobile/browser/iosbrowserdriver.py
@@ -1,23 +1,24 @@
 # -*- coding: utf-8 -*-
 from nzme_skynet.core.driver.mobile.browser.mbrowserdriver import MBrowserDriver
+import logging
+
+logger = logging.getLogger(__name__)
 
 
 class IOSBrowserDriver(MBrowserDriver):
-
-    def __init__(self, desired_capabilities, remote_url):
-        super(IOSBrowserDriver, self).__init__(desired_capabilities, remote_url)
 
     def _create_desired_capabilities(self):
         if not self._desired_cap:
             raise Exception("No capabilities provided to init Appium driver")
         if not self._desired_cap['platformVersion']:
             raise Exception("Please provide platformVersion to test app against")
-        if self._browser not in ['Safari', 'Chrome']:
+        if self._browser not in ['safari', 'chrome']:
             raise Exception("Only supports Safari and Chrome browsers on IOS")
 
-        self._desired_cap['browserName'] = self._browser
+        if "chrome" in self._desired_cap['browserName']:
+            # To disable the welcome screen on launching chrome
+            self.add_chrome_options('--no-first-run')
         self._desired_cap['platformName'] = 'iOS'
-        self._desired_cap['platform'] = 'iOS'
         if not self._desired_cap['deviceName']:
-            # Run tests on Android emulator by default
-            self._desired_cap['deviceName'] = 'iPhone Simulator'
+            raise Exception("No iOS deviceName specified")
+        logger.debug("iOS capability for creating driver {0}".format(self._desired_cap))