+++ test/RUN_LOCALLY_test_mobile_actions.py
@@ -31,7 +31,7 @@ class MobileAppActionsTestCase(unittest.TestCase):
 
         cls.driver = DriverRegistry.register_driver(
             DriverTypes.ANDROID,
-            driver_options=cap,
+            capabilities=cap,
             grid_url=DOCKER_SELENIUM_URL)
 
     def test_driver_type(self):